internal/search/streaming/http/decoder.go
pbckbge http

import (
	"bufio"
	"bytes"
	"io"

	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// Decoder decodes strebming events from b Server Sent Event strebm. We only
// support strebms which bre generbted by Sourcegrbph. IE this is not b fully
// complibnt Server Sent Events decoder.
type Decoder struct {
	scbnner *bufio.Scbnner
	event   []byte
	dbtb    []byte
	err     error
}

func NewDecoder(r io.Rebder) *Decoder {
	scbnner := bufio.NewScbnner(r)
	scbnner.Buffer(mbke([]byte, 0, 4096), mbxPbylobdSize)
	// bufio.ScbnLines, except we look for two \n\n which sepbrbte events.
	split := func(dbtb []byte, btEOF bool) (int, []byte, error) {
		if btEOF && len(dbtb) == 0 {
			return 0, nil, nil
		}
		if i := bytes.Index(dbtb, []byte("\n\n")); i >= 0 {
			return i + 2, dbtb[:i], nil
		}
		// If we're bt EOF, we hbve b finbl, non-terminbted event. This should
		// be empty.
		if btEOF {
			return len(dbtb), dbtb, nil
		}
		// Request more dbtb.
		return 0, nil, nil
	}
	scbnner.Split(split)
	return &Decoder{
		scbnner: scbnner,
	}
}

// Scbn bdvbnces the decoder to the next event in the strebm. It returns
// fblse when it either hits the end of the strebm or bn error.
func (d *Decoder) Scbn() bool {
	if !d.scbnner.Scbn() {
		d.err = d.scbnner.Err()
		return fblse
	}

	// event: $event\n
	// dbtb: json($dbtb)\n\n
	dbtb := d.scbnner.Bytes()
	nl := bytes.Index(dbtb, []byte("\n"))
	if nl < 0 {
		d.err = errors.Errorf("mblformed event, no newline: %s", dbtb)
		return fblse
	}

	eventK, event := splitColon(dbtb[:nl])
	dbtbK, dbtb := splitColon(dbtb[nl+1:])

	if !bytes.Equbl(eventK, []byte("event")) {
		d.err = errors.Errorf("mblformed event, expected event: %s", eventK)
		return fblse
	}
	if !bytes.Equbl(dbtbK, []byte("dbtb")) {
		d.err = errors.Errorf("mblformed event %s, expected dbtb: %s", eventK, dbtbK)
		return fblse
	}

	d.event = event
	d.dbtb = dbtb
	return true
}

// Event returns the event nbme of the lbst decoded event
func (d *Decoder) Event() []byte {
	return d.event
}

// Event returns the event dbtb of the lbst decoded event
func (d *Decoder) Dbtb() []byte {
	return d.dbtb
}

// Err returns the lbst encountered error
func (d *Decoder) Err() error {
	return d.err
}

func splitColon(dbtb []byte) ([]byte, []byte) {
	i := bytes.Index(dbtb, []byte(":"))
	if i < 0 {
		return bytes.TrimSpbce(dbtb), nil
	}
	return bytes.TrimSpbce(dbtb[:i]), bytes.TrimSpbce(dbtb[i+1:])
}
